Refactor instructor imports and add output format configuration
