+++ uglychain/llm/__init__.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 
 from .base import BaseLanguageModel
-from .instructor_yaml import Instructor, ParseError
+from .instructor import Instructor, ParseError
 from .tools import FunctionCall, finish, run_function
 
 __all__ = [
